chat-stream-hook: send heartbeat to viewers
To help prevent proxy servers timing out the connection. ~s30 is a bit
longer than eyre uses for SSE connections, but it doesn't seem to make
too big a difference.

Possible optimization here is not sending heartbeats when there are no
viewers, but we're punting on taking on that timer juggling complexity
for the time being.