fafa: add export & import functionality
Adds an action for writing state to a jamfile in .urb/put, as well as an
action for loading that state back into the agent. Import will fail if
it would overwrite any existing factors, just in case.

We notably do not let you download the jamfile through the browser, as
this would make the secret keys accessible to other agents, something
we have been trying to avoid.
Accessing .urb/put might be difficult for people using hosting services,
but if it's completely impossible, they need to find a better host.

Because we had not previously accounted for effectful POSTs, we have to
touch all of the product lines in that part of the agent. Clearly we
should just standardize on a maximally flexible version of this pattern.

Closes #8.