gossip: gracefully handle future gossip versions
Mark-based versioning is not an option, because we do want to hear the
gossip, even if we don't necessarily understand it. Instead, we opt for
an internal version indicator, and simply allowing any type for the
rumor metadata.

For versions we understand (currently just %0) we handle them
appropriately. For versions we don't understand, we queue them up, with
the expectation that, because some instance of this application is
already broadcasting it, we will in the (hopefully near) future obtain a
version of this application that uses an updated gossip library, capable
of understanding the yet-unknown kind of rumor.