bbcode: parsing & rendering
Adds parsing and rendering for most of the "basic" bbcode tags.

Parsing is permissive in that any bbcode that fails to parse get
interpreted as plaintext instead.

Rendering should probably (but arguably) be updated to wrap separate
paragraphs in actual <p> tags.

The [size] and [color] tags are vulnerable to arbitrary css injection.
This isn't the end of the world, but should be fixed soon.

Lightly tested, but not yet used in practice. Soon...