gossip: stop including timestamp with rumors
We now simply remember content by its hash.

The thinking here was that the timestamp could be used for relay loop
protection or something, but the content hash takes care of that by
itself. The timestamp ended up complicating things, because attaching it
in the initial subscription response was not accurate and led to not
recognizing duplicate content on-resubscribe.