sigil: make symbol map static
Instead of generating all symbols for the given colors, we now produce a
static map that contains functions for generating individual symbols
with given colors. This is only a tiny bit slower when generating with
the same color over and over, but two orders of magnitude faster when
generating for a wide variety of colors.