then: early, very rough prototype
This is the code as it has existed since Q4 2022. Committing now to
for the record, and to clearly demarcate the next phase of work,
which might change parts of this substantially.

As it stands, this has implementation-design issues primarily around
prefabs. Very uncertain about what they need to look like. An early,
rough, incomplete example is provided in /app/then/part-pals. It doesn't
work, and cries out for help.

$then's %make case, taking a plain @t for usercode, is probably a fine
fit for user-provided code, but is rather awkward for app-supplied
prefabs, which are the primary thing that we want to make good: most
users won't be writing their own code, but will want to use logic
provided by the apps they have installed, and use that to pipe together
workflows.

Flows in general already work, though their functionality is somewhat
limited right now.