gossip: add optional "proxying" logic
When the new "pass" setting is enabled, there's a 50% chance emitted
rumors will enter the proxying flow. Instead of sending the datum out as
a fact to all subscribers, we pick a peer at random, and poke them with
the datum. In the common case, they will in turn take a 50% chance on
whether to give the rumor as a fact, or pass it as a poke.

The aim of this is to provide a mode of operation where it's much more
difficult to ascertain the original creator of any given datum. Without
this new setting enabled, one can discern whether they got the datum
from its original source by simply looking at the hop counter. That same
counter could be used to narrow in on small subsets of ships, given a
broad enough malicious social graph.

Rumors we hear through pokes get processed by the underlying agent as
normal. They don't, however, get remembered as having been broadcast, so
that if we eventually receive it as a fact, we relay as normal. This
helps obscure whether we ever proxied it or not.

Datums with timestamps in them may still be slightly sensitive to
analysis, but this shouldn't be enough to identify the original
sender with any serious amount of certainty.