rumors: subtle drop shadow
This adds a drop shadow to rumors text that is light (15% black), diffuse (3px radius), and only slightly offset (1px x 1px). The net effect is that it makes the text more legible on light backgrounds.

Further, the text is spaced out an additional 0.1px, a subtle change that adds a bit of breathing room to the text.