then: broad refinements & early web ui
Builds upon the direction from the second prototype. Implements
de/inflation in +on-save and +on-load respectively, adding flow
(re)building logic that properly catches mismatching configuration
specs, build failures, and so on.

Includes a simple web ui that walks you, linearly, though constructing a
flow. Transformers and actions are filtered for compatibility,
configuration input fields are generated automatically. Everything about
this could be fancier, nicer, more legible... but it works!

Touched up the $yard-related types. Updated the pals yard to match, and
added some built-ins that %then might want to ship with. Including it as
%then's yard for the time being.