pals: add rudimentary breach handling
Pals now listens to jael's %public-keys endpoint. When it hears of a
breach for a ship that it knows, incoming status for that ship is
removed, and if we have them as a pal, we tell them again, hoping they
will hear the good news post-breach.

Additionally, pals now has a %resend command that may be used to
re-notify all known ships about our relation to them. Here, we use this
to partially resolve mismatched state caused by breaches prior to these
changes. To fully resolve those, we may need to run %resend again at a
later point in time.