million: support new chat backend
We detect whether the channels agent is running locally. If it's not, we
assume the local ship is still using the old chat backend, and use the
old %dm-action mark. Otherwise, we check to make sure that the new chat
backend is at chat-dm protocol version 0, and then send that a
%chat-dm-action instead.