scooore: update for latest iteration of channels
We do away with logic for old chat entirely, and now talk exclusively to
the channels agent. We negotiate versions with it, meaning our scries
shouldn't fail as long as we trigger them off subscription events. The
agent will safely sit and wait until the channels agent is running and
versions match.