scooore: use %noun instead of %channel-* marks
These mark names changed in the upcoming groups release, but the
underlying types remained the same. We don't actually care about
mark-level validation, we know the type, so this way it (probably) works
with both the old and the new api.

(Really we should move to doing scries at the new versioned endpoints,
but those are too new right now, we'll transition to using those once
they've been out in the wild for a little while, probably at a kelvin
version border.)