gossip: avoid known-bad relays
When the pass.config flag is set to true, we may relay gossip through
one of our peers. If this peer fails to acknowledge the relay in a
timely fashion, we retry elsewhere.

We had failed to realize that the timed out attempts aren't "free". We
pay for them in outgoing ames flows, and the peers we try relaying
through pay for them in incoming traffic from those flows. The common
failure case here, where the peer isn't running the agent we poked, is
costly in that this flow never resolves: their gall refuses to
acknowledge the message, and we're forced to keep retrying, giving both
of us increased load.

In some sense, all these pokes should happen on the same flow, so that
beyond the first attempt the load doesn't increase (as the first poke
would need to be resolved for the second one to be sent). However, we
want to track "dead" peers anyway (to increase the chance of doing a
successfuly relay on first try), and doing so gets us to the same
result, where after a single failed attempt we don't try again until we
actually hear back from them.

So we now track "dead" peers, peers who failed to relay in a timely
manner or nacked our relay request. We remove peers from the "misses"
whenever they ack our relay request, or whenever we hear gossip from
them (through fact or poke), because that indicates they are alive and
running a version of the agent/library that we can talk to.

It is, unfortunately, not possible for us to clean up the outstanding
pokes we've already created this way. (This issue became apparent after
~palfun-foslup slowed down due to large amount of attempts to poke a
gossiping agent that wasn't installed on it. Other ships likely suffer
from similar "dead-end traffic".) Best one can do without installing the
affected agent(s) is installing a skeleton agent in their place to
process the pokes so the queues clear out.