spots: support location sharing over ames
Adds a bunch of state for:
- other people's locations
- other people's display details (nickname, avatar)
- incoming location tracking requests
- location sharing permissions
- updates to send to our own devices

Adds a subscription endpoint, /live/[src], for getting all updates we
were granted access to. Should probably support something more granular
at some point, even if just for local-only usage.

Important detail is that for avatars, because we need to keep them in
state so that we can send them down to Owntracks (and because we can't
convert SVGs into PNGs on-urbit) we call out to a pal.dev "aide", a
microservice if you will, that downloads, downscales, png-izes and
compresses the avatar image for us.

(Those endpoints don't intentionally keep usage logs. The requests
aren't authenticated in any way, but they do obviously "leak" the
requester's IP address, and the avatar's url or @p. We consider this
"fine for now". If you have beef, replace with your own endpoint, or let
me know and I can make them hot-swappable. Should probably go to ~paldev
or some other ship in a "service provider" kind of way, but we punt on
that for the time being.)