owntracks: introduce +bu, for +ut's %r case
This makes the common usage of %r much nicer to write.

You might assume that %r should just take a bunt, but look at +ut and
come up with an ergonomic way of doing that. Maybe having the .wer type
be a proper tagged union wouldn't be so bad, but I'm not yet
convinced... Besides, who knows, we might need to flexibility of %r one
day!