spots: use transition events as location fallback
If a device's location data is stale (ie we haven't received a location
update from it in five minutes), and we receive a transition event for
it, we now try to turn that transition into a location update event.

We can only do this if the transition event contains coordinates, or
it's an "enter" event for a waypoint/zone that we know about.