spots: support "clubs", groups for guest users
Prototype implementation of a new feature that lets users with just the
Owntracks app (no ship of their own) co-locate on the host's %spots
instance.

Usernames for auth are first-come first-serve, the password provided is
static and cannot be changed. Users are identified by their device-id,
which we might suggest-set to be their guest comet id to do correlation
in the browser client.

We include the hoon-side page, but none of the js artefacts yet. Still
ironing out build scripts, file management, file server details etc.