owntracks: catch absent tid field case
In the app, if the tracker ID is not set at all, it also won't be sent
with location or waypoint messages.

Here, we simply fall back to the empty string, rather than unitizing the
tracker ID. This is less disruptive in the short term, but also matches
their docs and the expected behavior better. (Tracker ID really is a bad
ID field, anyway.)