foo-fileserver: generic from-clay file-serving
Intended to be copied into desks as a standalone %deskname-fileserver
agent. Only the configuration at the head of the file needs to be
changed (and can be changed after-the-fact).

Serves files under the specified web-root as they exist in clay under
the specified file-root. Is forced to do stupid in-place mark/to-mime
conversions that may crash, but that generally shouldn't happen if your
marks are any good.