spots: initial /web commit
Paldev has fallen...

Serving dynamic, interactive web frontends without using any external
tools is a grind. So far, we've managed to make do without any
dependencies at all. The share page, though primitive, was just a couple
lines of javascript. But it also didn't do much beyond polling for a
json file and updating a single map marker based on the result.

Here, we want to provide more interactivity. Listing multiple map
markers, changing focus on-tap, showing and hiding additional details,
etc.

We stubbornly stay as minimal as possible, but cannot avoid adding
dependencies. At the very least, we must insist on typescript, which
implies tsc. From there, we desire to bundle multiple files into one for
ease of loading. There's plenty of trade-offs being made, and we may
revisit them at some point.

Note that all the javascript here was hastily written with the intent of
getting a working prototype out the door as fast as possible. Approach
and thinking changes over the course of development, and as such the
whole feels completely unprincipled, or at least incoherent.

We continue to learn and evolve. Current thinking is we may hand-roll a
vanjs clone at some point. Writing it ourselves (instead of pulling in
vanjs as-is, even though it's just 140 lines) might enable us to
optimize its devex for working with @urbit's http-api and nockjs.

But who knows where we'll end up. Time will tell!