foo-fileserver: do tombstoning on file root
For any file root given to the fileserver, it will set a "do not keep"
tombstoning policy, and on every refresh issue a %tomb %pick to make
clay actually do garbage collection on the desk.

Careful, this means that any directory served by the fileserver will not
retain clay history! (Maybe we should have a config flag for this...)